// mobile/src/lib/types.ts
export interface Category {
  id: number;
  name: string;
  slug: string;
  parent_id: number | null;
  products_count: number;
}
// ...
export interface CategoryInput {
  name: string;
  parent_id?: number | null;
  slug?: string;
}